# contracts/generated/python/property_playground_contracts/models/model_info.py
# coding: utf-8

"""
    Property Playground Model API

    Canonical contract for housing feature validation and price prediction across the Python services, Java market API, and Next.js portal.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, StrictFloat, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Union
from property_playground_contracts.models.model_metrics import ModelMetrics
from typing import Optional, Set
from typing_extensions import Self

class ModelInfo(BaseModel):
    """
    ModelInfo
    """ # noqa: E501
    model_type: StrictStr
    model_version: StrictStr
    feature_names: List[StrictStr]
    coefficients: Dict[str, Union[StrictFloat, StrictInt]]
    intercept: Union[StrictFloat, StrictInt]
    metrics: ModelMetrics
    training_source: StrictStr
    __properties: ClassVar[List[str]] = ["model_type", "model_version", "feature_names", "coefficients", "intercept", "metrics", "training_source"]

    model_config = ConfigDict(
        populate_by_name=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of ModelInfo from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of metrics
        if self.metrics:
            _dict['metrics'] = self.metrics.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of ModelInfo from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "model_type": obj.get("model_type"),
            "model_version": obj.get("model_version"),
            "feature_names": obj.get("feature_names"),
            "coefficients": obj.get("coefficients"),
            "intercept": obj.get("intercept"),
            "metrics": ModelMetrics.from_dict(obj["metrics"]) if obj.get("metrics") is not None else None,
            "training_source": obj.get("training_source")
        })
        return _obj