# portal/lib/generated/openapi/models/ModelInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Property Playground Model API
 * Canonical contract for housing feature validation and price prediction across the Python services, Java market API, and Next.js portal.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ModelMetrics } from './ModelMetrics';
import {
    ModelMetricsFromJSON,
    ModelMetricsFromJSONTyped,
    ModelMetricsToJSON,
    ModelMetricsToJSONTyped,
} from './ModelMetrics';

/**
 *
 * @export
 * @interface ModelInfo
 */
export interface ModelInfo {
    /**
     *
     * @type {string}
     * @memberof ModelInfo
     */
    model_type: string;
    /**
     *
     * @type {string}
     * @memberof ModelInfo
     */
    model_version: string;
    /**
     *
     * @type {Array<string>}
     * @memberof ModelInfo
     */
    feature_names: Array<string>;
    /**
     *
     * @type {{ [key: string]: number; }}
     * @memberof ModelInfo
     */
    coefficients: { [key: string]: number; };
    /**
     *
     * @type {number}
     * @memberof ModelInfo
     */
    intercept: number;
    /**
     *
     * @type {ModelMetrics}
     * @memberof ModelInfo
     */
    metrics: ModelMetrics;
    /**
     *
     * @type {string}
     * @memberof ModelInfo
     */
    training_source: string;
}

/**
 * Check if a given object implements the ModelInfo interface.
 */
export function instanceOfModelInfo(value: object): value is ModelInfo {
    if (!('model_type' in value) || value['model_type'] === undefined) return false;
    if (!('model_version' in value) || value['model_version'] === undefined) return false;
    if (!('feature_names' in value) || value['feature_names'] === undefined) return false;
    if (!('coefficients' in value) || value['coefficients'] === undefined) return false;
    if (!('intercept' in value) || value['intercept'] === undefined) return false;
    if (!('metrics' in value) || value['metrics'] === undefined) return false;
    if (!('training_source' in value) || value['training_source'] === undefined) return false;
    return true;
}

export function ModelInfoFromJSON(json: any): ModelInfo {
    return ModelInfoFromJSONTyped(json, false);
}

export function ModelInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelInfo {
    if (json == null) {
        return json;
    }
    return {

        'model_type': json['model_type'],
        'model_version': json['model_version'],
        'feature_names': json['feature_names'],
        'coefficients': json['coefficients'],
        'intercept': json['intercept'],
        'metrics': ModelMetricsFromJSON(json['metrics']),
        'training_source': json['training_source'],
    };
}

export function ModelInfoToJSON(json: any): ModelInfo {
    return ModelInfoToJSONTyped(json, false);
}

export function ModelInfoToJSONTyped(value?: ModelInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'model_type': value['model_type'],
        'model_version': value['model_version'],
        'feature_names': value['feature_names'],
        'coefficients': value['coefficients'],
        'intercept': value['intercept'],
        'metrics': ModelMetricsToJSON(value['metrics']),
        'training_source': value['training_source'],
    };
}